+++ common/include/common/GameConfig.h
@@ -54,6 +54,7 @@ struct GameConfig
     bool allow_overwrite_game_files = false;
     bool scoreboard_anim = false;
     bool keep_launcher_open = false;
+    int skip_cutscene_ctrl = -1;
 
     // Advanced
     float level_sound_volume = 1.0f;